Make sure the initial empty `SynapticPathway.spiking_synapses` attribute has the correct dtype

+++ brian2/synapses/synapses.py
@@ -152,7 +152,7 @@ class SynapticPathway(CodeRunner, Group):
         self.spikes_start = self.source.start
         self.spikes_stop = self.source.stop
 
-        self.spiking_synapses = np.array([], dtype=int)
+        self.spiking_synapses = np.array([], dtype=np.int32)
         self.variables = Variables(self)
         self.variables.add_attribute_variable('_spiking_synapses', unit=Unit(1),
                                               obj=self,
